copy shared code into each component
