+++ starter/src/registry/hs1-albany-new-patients/components/Hs1AlbanyNewPatientsBreadcrumbsSection.tsx
@@ -1,5 +1,6 @@
+import type { ComponentProps } from "react";
+import { Link as PagesLink } from "@yext/pages-components";
 import { ComponentConfig, Fields, PuckComponent } from "@puckeditor/core";
-import { Link } from "../../shared/SafeLink";
 import {
   resolveComponentData,
   TranslatableString,
@@ -7,6 +8,30 @@ import {
   YextEntityField,
   YextEntityFieldSelector,
 } from "@yext/visual-editor";
+
+const getSafeHref = (href?: string): string => {
+  const trimmedHref = href?.trim();
+  return trimmedHref ? trimmedHref : "#";
+};
+
+type PagesLinkProps = ComponentProps<typeof PagesLink>;
+
+const Link = (props: PagesLinkProps) => {
+  const safeProps = { ...props } as any;
+
+  if ("cta" in safeProps && safeProps.cta) {
+    safeProps.cta = {
+      ...safeProps.cta,
+      link: getSafeHref(safeProps.cta.link),
+    };
+  }
+
+  if ("href" in safeProps) {
+    safeProps.href = getSafeHref(safeProps.href);
+  }
+
+  return <PagesLink {...safeProps} />;
+};
 
 type StyledTextProps = {
   text: YextEntityField<TranslatableString>;
