+++ starter/src/registry/hs1-albany-new-patients/components/Hs1AlbanyNewPatientsHeaderSection.tsx
@@ -1,6 +1,7 @@
+import type { ComponentProps } from "react";
+import { Link as PagesLink } from "@yext/pages-components";
 import { ComponentConfig, Fields, PuckComponent } from "@puckeditor/core";
 import { ComplexImageType, ImageType } from "@yext/pages-components";
-import { Link } from "../../shared/SafeLink";
 import {
   Image,
   resolveComponentData,
@@ -11,6 +12,30 @@ import {
   YextEntityFieldSelector,
 } from "@yext/visual-editor";
 import { useState } from "react";
+
+const getSafeHref = (href?: string): string => {
+  const trimmedHref = href?.trim();
+  return trimmedHref ? trimmedHref : "#";
+};
+
+type PagesLinkProps = ComponentProps<typeof PagesLink>;
+
+const Link = (props: PagesLinkProps) => {
+  const safeProps = { ...props } as any;
+
+  if ("cta" in safeProps && safeProps.cta) {
+    safeProps.cta = {
+      ...safeProps.cta,
+      link: getSafeHref(safeProps.cta.link),
+    };
+  }
+
+  if ("href" in safeProps) {
+    safeProps.href = getSafeHref(safeProps.href);
+  }
+
+  return <PagesLink {...safeProps} />;
+};
 
 type StyledTextProps = {
   text: YextEntityField<TranslatableString>;
