+++ starter/src/registry/hs1-albany-office/components/Hs1AlbanyOfficeLocationSection.tsx
@@ -1,3 +1,6 @@
+import type { ComponentProps } from "react";
+import { Address as PagesAddress } from "@yext/pages-components";
+import { Link as PagesLink } from "@yext/pages-components";
 import { ComponentConfig, Fields, PuckComponent } from "@puckeditor/core";
 import {
   resolveComponentData,
@@ -6,8 +9,55 @@ import {
   YextEntityField,
   YextEntityFieldSelector,
 } from "@yext/visual-editor";
-import { Address } from "../../shared/SafeAddress";
-import { Link } from "../../shared/SafeLink";
+
+type PagesAddressProps = ComponentProps<typeof PagesAddress>;
+
+const hasAddressContent = (address: PagesAddressProps["address"]): boolean => {
+  if (!address) {
+    return false;
+  }
+
+  return Boolean(
+    address.line1 ||
+      address.line2 ||
+      address.city ||
+      address.region ||
+      address.postalCode ||
+      address.countryCode,
+  );
+};
+
+const Address = (props: PagesAddressProps) => {
+  if (!hasAddressContent(props.address)) {
+    return null;
+  }
+
+  return <PagesAddress {...props} />;
+};
+
+const getSafeHref = (href?: string): string => {
+  const trimmedHref = href?.trim();
+  return trimmedHref ? trimmedHref : "#";
+};
+
+type PagesLinkProps = ComponentProps<typeof PagesLink>;
+
+const Link = (props: PagesLinkProps) => {
+  const safeProps = { ...props } as any;
+
+  if ("cta" in safeProps && safeProps.cta) {
+    safeProps.cta = {
+      ...safeProps.cta,
+      link: getSafeHref(safeProps.cta.link),
+    };
+  }
+
+  if ("href" in safeProps) {
+    safeProps.href = getSafeHref(safeProps.href);
+  }
+
+  return <PagesLink {...safeProps} />;
+};
 
 type StyledTextProps = {
   text: YextEntityField<TranslatableString>;
