+++ starter/src/registry/hs1-chicago/components/Hs1ChicagoLocationHoursSection.tsx
@@ -1,3 +1,5 @@
+import type { ComponentProps } from "react";
+import { Address as PagesAddress } from "@yext/pages-components";
 import { ComponentConfig, Fields, PuckComponent } from "@puckeditor/core";
 import {
   resolveComponentData,
@@ -7,7 +9,31 @@ import {
   YextEntityFieldSelector,
 } from "@yext/visual-editor";
 import { HoursTable } from "@yext/pages-components";
-import { Address } from "../../shared/SafeAddress";
+
+type PagesAddressProps = ComponentProps<typeof PagesAddress>;
+
+const hasAddressContent = (address: PagesAddressProps["address"]): boolean => {
+  if (!address) {
+    return false;
+  }
+
+  return Boolean(
+    address.line1 ||
+      address.line2 ||
+      address.city ||
+      address.region ||
+      address.postalCode ||
+      address.countryCode,
+  );
+};
+
+const Address = (props: PagesAddressProps) => {
+  if (!hasAddressContent(props.address)) {
+    return null;
+  }
+
+  return <PagesAddress {...props} />;
+};
 
 type StyledTextProps = {
   text: YextEntityField<TranslatableString>;
