+++ starter/src/registry/hs1-laguna/components/Hs1LagunaFeaturedArticlesSection.tsx
@@ -1,6 +1,31 @@
+import type { ComponentProps } from "react";
+import { Link as PagesLink } from "@yext/pages-components";
 import { useState } from "react";
 import { ComponentConfig, Fields, PuckComponent } from "@puckeditor/core";
-import { Link } from "../../shared/SafeLink";
+
+const getSafeHref = (href?: string): string => {
+  const trimmedHref = href?.trim();
+  return trimmedHref ? trimmedHref : "#";
+};
+
+type PagesLinkProps = ComponentProps<typeof PagesLink>;
+
+const Link = (props: PagesLinkProps) => {
+  const safeProps = { ...props } as any;
+
+  if ("cta" in safeProps && safeProps.cta) {
+    safeProps.cta = {
+      ...safeProps.cta,
+      link: getSafeHref(safeProps.cta.link),
+    };
+  }
+
+  if ("href" in safeProps) {
+    safeProps.href = getSafeHref(safeProps.href);
+  }
+
+  return <PagesLink {...safeProps} />;
+};
 
 type ArticleItem = {
   title: string;
