--- starter/src/registry/shared/SafeLink.tsx
@@ -1,26 +1,0 @@
-import type { ComponentProps } from "react";
-import { Link as PagesLink } from "@yext/pages-components";
-
-const getSafeHref = (href?: string): string => {
-  const trimmedHref = href?.trim();
-  return trimmedHref ? trimmedHref : "#";
-};
-
-type PagesLinkProps = ComponentProps<typeof PagesLink>;
-
-export const Link = (props: PagesLinkProps) => {
-  const safeProps = { ...props } as any;
-
-  if ("cta" in safeProps && safeProps.cta) {
-    safeProps.cta = {
-      ...safeProps.cta,
-      link: getSafeHref(safeProps.cta.link),
-    };
-  }
-
-  if ("href" in safeProps) {
-    safeProps.href = getSafeHref(safeProps.href);
-  }
-
-  return <PagesLink {...safeProps} />;
-};
